Implement persistent search box in search results page
This was a really tricky one.

The challenge was to get to know when the user is searching a new search term

This was easy on the homepage because we were loading the Search Results Component and we have the ngOnInit hook to call the search function

But once we have loaded the search results, when the user searches for something else, we cannot use the ngOnInit hook because the component is already loaded.

I noticed that the URL was changing ecerytime we called the search. So i subscribed to the ActivatedRoute objects' paramMap property and called the search function everytime it changed.

It was very satisfying to see this functionality completed after moving through the fog of code online at 2 in the night.